requirements.txt
fastapi
uvicorn
pandas
scikit-learn
tensorflow
numpy
opencv-python
pytesseract
Pillow
ortools
nltk
python-multipart
requests
imagehash
firebase-admin
google-cloud-firestore
python-dotenv
